requirements.txt
torch
transformers
accelerate
numpy
pandas
pillow
python-dotenv
huggingface_hub>=0.24.0
faiss-cpu
tqdm
requests
scikit-learn
